Finished-Question-5-responds with all articles
