+++ app.js
@@ -1,4 +1,6 @@
-const { getAllTopics, getEndPoints, getArticlesById }= require('./controllers/controller')
+const { getArticlesById, getAllArticles  } = require('./controllers/articles.controller');
+const { getAllTopics } = require('./controllers/topics.controller')
+const { getEndPoints } = require('./controllers/api.controller') 
 const express = require('express');
 const app = express();
 
@@ -9,6 +11,9 @@ app.get('/api/topics', getAllTopics );
 app.get('/api', getEndPoints);
 
 app.get('/api/articles/:article_id',getArticlesById )
+
+app.get('/api/articles', getAllArticles)
+
 
 app.use((err,req,res,next) => {
     console.log(err);
