+++ endpoints.json
@@ -29,16 +29,15 @@
   "GET /api/articles/:articles_id": {
     "description": "serves an array of articles by id",
     "queries": [],
-    "exampleResponse":{
-    "title": "Running a Node App",
-    "topic": "coding",
-    "author": "jessjelly",
-    "body": "This is part two of a series on how to get up and running with Systemd and Node.js. This part dives deeper into how to successfully run your app with systemd long-term, and how to set it up in a production environment.",
-    "created_at": 1604728980000,
-    "votes": 0,
-    "article_img_url":
-      "https://images.pexels.com/photos/11035380/pexels-photo-11035380.jpeg?w=700&h=700"
+    "exampleResponse": {
+      "title": "Running a Node App",
+      "topic": "coding",
+      "author": "jessjelly",
+      "body": "This is part two of a series on how to get up and running with Systemd and Node.js. This part dives deeper into how to successfully run your app with systemd long-term, and how to set it up in a production environment.",
+      "created_at": 1604728980000,
+      "votes": 0,
+      "article_img_url":
+        "https://images.pexels.com/photos/11035380/pexels-photo-11035380.jpeg?w=700&h=700"
     }
-
   }
 }
